+++ lib/src/builder.rs
@@ -1,6 +1,8 @@
 use model::{Body, NotesMetadata};
 use util::generate_uuid;
-use note::RemoteNoteMetaData;
+use note::{RemoteNoteMetaData, NoteHeaders, HeaderParser};
+use chrono::Utc;
+use profile;
 
 pub struct NoteTupleBuilder {
     metadata: NotesMetadata,
@@ -67,6 +69,7 @@ pub struct NotesMetadataBuilder {
 /// generated
 impl NotesMetadataBuilder {
     pub fn new() -> NotesMetadataBuilder {
+        let date = Utc::now().to_rfc2822();
         NotesMetadataBuilder {
             notesMetadata:  NotesMetadata {
                 old_remote_id: None,
@@ -74,9 +77,9 @@ impl NotesMetadataBuilder {
                 locally_deleted: false,
                 locally_edited: false,
                 new: false,
-                date: "".to_string(),
+                date,
                 uuid: generate_uuid(),
-                mime_version: "".to_string()
+                mime_version: "1.0 (Mac OS X Notes 4.6 \\(879.10\\))".to_string()
             }
         }
 
@@ -112,5 +115,59 @@ impl NotesMetadataBuilder {
             folder: "".to_string(),
             uid: 0
         }
+    }
+}
+
+pub struct HeaderBuilder {
+    headers: Vec<(String,String)>
+}
+
+impl HeaderBuilder {
+
+    pub fn new() -> HeaderBuilder {
+        let mut headers: Vec<(String,String)> = vec![];
+        let profile = profile::load_profile();
+        headers.push(("X-Uniform-Type-Identifier".to_string(), "com.apple.mail-note".to_string()));
+        headers.push(("Content-Type".to_string(), "text/html; charset=utf-8".to_string()));
+        headers.push(("Content-Transfer-Encoding".to_string(), "quoted-printable".to_string()));
+        headers.push(("Mime-Version".to_string(), "1.0 (Mac OS X Notes 4.6 \\(879.10\\))".to_string()));
+        let date = Utc::now().to_rfc2822();
+        headers.push(("Date".to_string(), date.clone()));
+        headers.push(("X-Mail-Created-Date".to_string(), date.clone()));
+        headers.push(("From".to_string(), profile.email));
+
+        HeaderBuilder {
+            headers
+        }
+    }
+
+    pub fn with_message_id(mut self, message_id: String) -> Self {
+        let profile = self::profile::load_profile();
+        self.headers.push(("Message-Id".to_string(), format!("<{}@{}", message_id, profile.domain())));
+        self
+    }
+
+    pub fn with_uuid(mut self, uuid: String) -> Self {
+        self.headers.push(("X-Universally-Unique-Identifier".to_string(), uuid));
+        self
+    }
+
+    pub fn with_subject(mut self, subject: &str) -> Self {
+        self.headers.push(("Subject".to_string(), subject.to_string()));
+        self
+    }
+
+    pub fn build(mut self) -> NoteHeaders {
+        let profile = self::profile::load_profile();
+
+        if None == self.headers.get_header_value("X-Universally-Unique-Identifier") {
+            self.headers.push(("X-Universally-Unique-Identifier".to_string(), generate_uuid()));
+        }
+
+        if None == self.headers.get_header_value("Message-Id") {
+            self.headers.push(("Message-Id".to_string(), format!("<{}@{}", generate_uuid(), profile.domain())));
+        }
+
+        self.headers
     }
 }
